Better UninitializedLocalCheck: Represent bottom better in `show`
Summary:
I'm looking into why UninitializedLocalCheck does not recognize various
branching cases where we don't have to worry about uninitialized locals.

One source of confusion is that the way we are printing Bottom state
makes it look as though bottom is equivalent to no locals existing, but
that's actually the opposite of what is going on: empty locals (which
is directly representanble as `Value Identifier.Map.empty`) is the
Top of our lattice, whereas `Bottom` represents unreachable code
and is treated in the lattice "as if" all possible locals are defined.

So we really don't want to implement `show` as returning `"[]"`, that
is extremely confusing.

Reviewed By: arthaud

Differential Revision: D37534285

fbshipit-source-id: bbaa8838e8383d6379df581ae7af4f08965f9b5a