Reorganize tests
Summary:
Several test cases slipped into the override decorator test
did not really belong there.

Here, I move each of them to where they belong (several are
actually postcondition tests, and one of them validates some
new "attribute kind" checks we added that verify we don't
override normal attributes with bound methods). I also remove
the extranious `override` decorators from tests where it is
irrelevant.

The one test case verifying that override attribute kind
is consistent (i.e. that we do not override plain attributes
with bound methods) belongs in a separate test function, because
it is not tied to the implementation of `override`.

Several of the override test cases were validating logic that didn't
have anything to do with the override decorator; previous commits
moved these test cases to where they belong, here we strip out the
decorator (which could mislead someone about what's actually being
tested).

I also filed T124880784 because I believe we have a soundness
hole in our handling of inheritance diamonds: a subtype in a diamond
has to be a valid instance of *all* its ancestors; method resolution order
does not have a role to play in this decision. But that's a preexisting
problem and unrelated to the oncall issue I am working on.

Reviewed By: shannonzhu

Differential Revision: D37536888

fbshipit-source-id: adcef2574c667ecf628a5c25e65a958b11002c17