Reproduce bug where we alias a class to a non-existent symbol.
Summary:
Problem: When a class is named the same as its module, and there is a particular combination of module imports, we end up aliasing the class to a non-existent symbol.

This took a *while* to pin down.

I'm not too familiar with Python's import resolution. Don't know the point at which we went wrong. Will repro and leave it as such.

Reviewed By: stroxler

Differential Revision: D36919168

fbshipit-source-id: 43d7ac98e5680048f1065571257b8eec9ed38815