Implement EnvironmentTable.Overlay
Summary:
The implementation is very similar to UnannotatedGlobalEnvironment.Overlay,
but even simpler:
- For update we just filter triggered keys with an ownership check before
  passing to the `FromReadOnlyUpstream` update code. Other than the ownership
  filters, updates work exactly like they do for `Base`.
- For read-only, we just wrap the parent `get` and this-read-only `get` with an ownership
  check.

The one wrinkle is that ownership checks are nontrivial for ParseAnnotationCache.
I think we can get away with a null ownership check (i.e. an environment lazily populates
*all* parsed annotations) but this is a possible perf issue; we could probably implement
some (possibly imprecise) ownership checks if it proves to be a problem.

Reviewed By: pradeep90

Differential Revision: D37387438

fbshipit-source-id: 50da956f85d615f6b1efadac9bf926f5e2ef7c4c