move validate_paths to Configuration.Analysis
Summary:
The validate paths functionality that has been historically living
in `check.ml` doesn't really belong there - we ought to be able to
validate paths, if we wish, even if we want to do something that
doesn't involve performing a full type check.

This is important for making a lazy server - which is something
we want both medium-term (to support IDE code navigation) and
almost immediately (to support using `pyre query` to dump an
inheritance graph without waiting for a pointless type check).

Reviewed By: grievejia

Differential Revision: D37583340

fbshipit-source-id: 8850a373330c5907e62c5432dc124ad6d1dde627