Move config-based filtering tests to ErrorsEnvironmentTest
Summary:
I'm about to remove the `Check` module, which doesn't really do
enough to be worth keeping these days - it is a very thin wrapper
around operations that belong in ErrorsEnvironment (and have been
moved).

The last blocker on removal is the test of error filtering configuration
flags in checkTest.ml; this commit moves those tests to
errorsEnvironmentTest.ml.

Reviewed By: grievejia

Differential Revision: D37623097

fbshipit-source-id: 96f1a9f907e3c9c5d471ab82590241a60beb42df