Strip `check.ml` down to two operations
Summary:
I've been cleaning up things we don't really need in check.ml for
quite a few commits now, with an eye toward understanding what it
should look like to allow lazy server initialization, which is something
we want for both pysa use cases (model generation, not the actual
taint analysis) and for a code navigation server.

At this point it has become clear that we don't really need check.ml
at all anymore, because after removing unneeded things and packaging
some of the production logging into "smarter' creation functions
we are really just doing 3 things:

- sanity checking the config
- creating a lazy environment
- eagerly populating that environment, which for now we do in
  2 steps just because that's what we've tuned the mapreduce
  parameters for (we can probably get rid of this if we are careful!)

This commit makes this clear by extracting the second and third
operations into `ErrorsEnvironment` functions, which is a better
place for them to live if we want to be able to mix and match more
flexibly.

Reviewed By: grievejia

Differential Revision: D37583341

fbshipit-source-id: 9b3dee2e5ecc4f8c6a3a0732d342584d12294b97