Get rid of the dotty integration
Summary:
We don't use this anymore, and its presence makes `check.ml` harder
to work with.

Let's just get rid of it - we're not losing anything fundamental, it's
still quite easy to call ClassHierarchy.to_dot in debug logs or even
hook it up to PyreQuery if we want to in the future.

Reviewed By: grievejia

Differential Revision: D37582515

fbshipit-source-id: f77ff9b0fc84e8103b4415d0bc9add1784178183