Parent Propagation: add propagation to UGE
Summary:
This commit adds parent update result propagation to UGE (this is
the base case of the general propagation operation, which is defined
recursively like all update operations - UGE is the base case
because it's the first environment layer with cross-module interactions).

The implementation is to:
- update the overlay by filtering down dependencies and then triggering
  re-computatin based on the parent's upstream (i.e. AstEnvironment)
  update result
- produce a combined UpdateResult where the triggered dependencies at the
  UGE layer is the union of dependencies coming from the parent and from the
  overlay; all subsequent layers are going to see this union when they look
  for what computation needs to be triggered.

Why do we need the union? Consider an example:
- My overlay has module `foo`; module `bar` is on disk
- module `bar` initially has two classes, `A` and `B`, and
  defines the type alias `Baz` to point at A.
- I have a wildcard import of module `bar` in module `foo`,
  and I make use of `Baz` as a parent class in class `Foo`.
- I rebase on a change that flips `Baz` from type `A` to type
  `B`

In this situation, any later stage of our overlay environment
(for example ClassHierarchyEnvironment) needs to know about *both*
the updates to `bar` on disk (for example did a new field get
added to `B` when we rebased?) as well as the change to the
preprocessed AST of `foo`. This is why we need a combined update
result that reflects both sets of triggered dependencies.

Reviewed By: grievejia

Differential Revision: D37492730

fbshipit-source-id: a9c680209a0cbbff2410a65620b013494db1daec