support for didClose: add a new operation to reset artifact path
Summary:
This commit uses the new type `ModuleTracker.Overlay.CodeUpdate.t` to
create the support we'll need for `didClose` language server events by
adding a `ResetCode` variant that tells the overlay to remove the
overlaid code.

This operation will not cause the artifact path to be removed from
the overlay (that's something we don't know how to do efficiently in
shared memory unfortunately - we'll need to wait on ocaml 5), but what
it does do is causes the overlay to read the code from disk. So, for example,
if the overlay is just one file then it will exactly match the parent
after performing this update.

Reviewed By: grievejia

Differential Revision: D37523233

fbshipit-source-id: 96a25edcd54cc17132df05c7c24b26f1f2a6639e