Parent Propagation: refactor UGE update code
Summary:
Now that we have overlay code updates working end-to-end with MultiEnvironment,
it's time to start looking at the other overlay operations we need:
- propagating parent updates (so that watchman changes reach the overlay)
- resetting an overlay to consume code from disk (for didClose events)

We'll start with the first of these because it is the trickier one and we
want to uncover any architectural issues as early as possible.

Parent propagation refers to the process of pushing incremental updates
across the boundary from parent environment to child - for example, if we
rebase and the definition of a class `Foo` defined in a module on disk changes,
the overlay for any open buffer that uses `Foo` needs to see this change.

This is something we only have to worry about for envirionment
layers that are affected by cross-module lookups, which means that the process
begins at the `UnannotatedGlobalEnvironment` layer - prior to that, all stored
data is 1:1 with raw code and therefore there's no way for a parent dependency
to trigger an overlay update.

The implementation will look like this:
- At most layers (the recursive case), we need
  - pass the upstream UpdaeResult from the parent environment to our own
    upstream layer. This will produce a new upstream UpdateResult
  - consume that new upstream UpdateResult, in exactly the same way we already
    already do in `update_overlaid_code`, which will give us some triggered dependencies
  - return a new UpdateResult in which
    - the upstream UpdateResult is what we got from our upstream (not the parent!)
    - the locally_triggered_dependencies is a concatenation of
      what we got from the parent and what we triggered ourselves.
  resulting UpdateResult
- At the UGE layer, (the base case) we need to be able to consume an
  AstEnvironment.UpdateResult coming from the *parent* environment. We can
  consume it using exactly the same
  logic we already use

This commit prepares for that bottom, base case layer by factorign the code to
consume an UpdateResult into a separate (private) function and having
`update_overlaid_code` just call that function on the output of
`AstEnvironment.Overlay.update_overlaid_code`.

I'll add the parent propagation function in the next commit.

Reviewed By: grievejia

Differential Revision: D37492732

fbshipit-source-id: fe2a64412a69b4dc7da0adf5ddb4dedff689877b