support for didClose: add CodeUpdate.t type
Summary:
In order to fully support unsaved editor changes, the overlay environments
need to handle three operations:
- new overlaid code (done): this is to support didChange events
- parent updates (done): this is to support watchman updates (e.g. on rebase)
  where the on-disk code changes and the in-editor state powered by the overlay
  needs to know about this
- resetting an overlay to the filesystem: this is to support didClose events

In the current sharedMemory-based environment codebase, it's not really possible
to remove an overlay when the editor closes; this will change whenever we are
able to move to ocaml 5 but that's a long way off. In the meantime, we're going
to need to handle didClose events in pretty much the same way as didChange events,
except that instead of updating the overlaid code in ModuleTracker.Overlay.t we
need to remove a key (so that the overlay still exists, but it gets its AST
from the filesystem; at this point the overlay state and root environment state
should generally match).

This commit prepares us for supporting close events by creating a new type,
`ModuleTracker.Overlay.CodeUpdate.t`, representing the update operation we
need to perform when calling `update_overlaid_code`. For now, the only update
is `NewCode` and it works just as before but in a follow-up commit we'll
add a new variant for resetting an artifact path to read from the parent
environment.

Reviewed By: grievejia

Differential Revision: D37523234

fbshipit-source-id: 4a084ee8d66468386bb370c3731c45140246e934