Blank response if expression level coverage enabled
Summary:
### Context:
We want to surface expression level coverage gaps through VSCode highlighting.

### Problem:
In this diff, we want the `--enable-expression-level-coverage` flag to return a blank dummy response for highlighting, else if there is no flag the original function level highlighting should persist.

### Solution:
If the flag is true, we return the dummy response. Else, we defer to existing behavior.

### Stack diff:
- Add dummy response for flag
- Refactor percentage so we can use expression level percentage instead of function level percentage when necessary
- Make diagnostics given a `ExpressionLevelCoverageResponse`
- Given path, give a expression level coverage response
- When querying for expression level coverage, give an expression level coverage response

Next: Refactor percentage

Reviewed By: stroxler

Differential Revision: D37648151

fbshipit-source-id: f0be6240a72077f596cf2df3cff4a09534bd77d4