Make diagnostics for expression level coverage
Summary:
### Context:
We want to surface expression level coverage gaps through VSCode highlighting.

### Problem:
In this diff, we want to make lsp diagnostics given an `ExpressionLevelCoverageResponse`.

### Solution:
To obtain VSCode diagnostics for a given path, we look at the first response in the `ExpressionLevelCoverageResponse` to see if there is coverage information. If there is coverage information, we turn each coverage gap into a diagnostic by making each `Location` into a `lsp.Range` with a corresponding message.

### Stack diff:
- Add dummy response for flag
- Refactor percentage so we can use expression level percentage instead of function level percentage when necessary
- Make diagnostics given a `ExpressionLevelCoverageResponse`
- Given path, give a expression level coverage response
- When querying for expression level coverage, give an expression level coverage response

Next: Given path, give a expression level coverage response

Reviewed By: pradeep90

Differential Revision: D37652719

fbshipit-source-id: 0f403bda1bd673cb045d1a1ab89b2d57df9e6f5c