Display VSCode diagnostics for expression level coverage
Summary:
### Context:
We want to surface expression level coverage gaps through VSCode highlighting.

### Problem:
In this diff, we want display highlights and percentage for expression level coverage when the flag is on.

### Solution:
When the flag for expression level coverage is on, we display highlights and percentage for expression level instead of function level coverage. Given the path, we query for the `ExpressionLevelCoverageResponse`. With this response, we use functions from the previous diffs in the stack to obtain diagnostics and percentages to surface.

### Stack diff:
- Add dummy response for flag
- Refactor percentage so we can use expression level percentage instead of function level percentage when necessary
- Make diagnostics given a `ExpressionLevelCoverageResponse`
- Given path, give a expression level coverage response
- When querying for expression level coverage, give an expression level coverage response

Reviewed By: pradeep90

Differential Revision: D37654255

fbshipit-source-id: d3c867ec1824136b3c324e2aa0e0c78abe182b92