Refactor percentage
Summary:
### Context:
We want to surface expression level coverage gaps through VSCode highlighting.

### Problem:
In this diff, we pull out the expression level percent coverage calculation so that we can later surface this in lieu of the function level percent calculation.

### Solution:
We pull out this calculation into a separate function

### Stack diff:
- Add dummy response for flag
- Refactor percentage so we can use expression level percentage instead of function level percentage when necessary
- Make diagnostics given a `ExpressionLevelCoverageResponse`
- Given path, give a expression level coverage response
- When querying for expression level coverage, give an expression level coverage response

Next: Make diagnostics given a `ExpressionLevelCoverageResponse`

Reviewed By: stroxler

Differential Revision: D37649625

fbshipit-source-id: ef0216a982005cf1b0209227c97009eced09a1fb