Refactor query tests for easier future modification
Summary:
### Context:
Currently, I'm adding coverage gap behavior which touches the query tests; however, the query tests are quite repetitive and hence difficult to debug when they break.

### Problem:
We want to refactor to remove the repetition, so that it is more readable and easier to work with.

### Solution:
The response for the file with a `#pyre-strict` is the same, with slight differences in path and function name, so we make this a separate function. We make a function for an error response and functions for building different types of responses for our test to use.

Reviewed By: stroxler

Differential Revision: D37663480

fbshipit-source-id: 913ef982451f62e156b980a25726d9b766486859