Lazy server for pysa: enable using the new flag in PyreConnection
Summary:
This commit finally wires up the PyreConnection so that in
programmatic use of `tools.api.pyre.query` it will be possible to make
use of a lazy server mode.

Since we want to rewrite the query API to be quite different in the
medium-term, I decided to take some shortcuts here. A few aspects of
this behavior aren't great:
- we return dummy type check results even when we're using lazy
  mode. This isn't a problem per se but suggests the design isn't great.
- in nonlazy mode, `start_server` will call `pyre incremental`, which
  will reuse an existing server. But in lazy mode calling start twice
  will result in an error unless some other call stopped the server
  first.
- for now, the restart_server api does not support lazy mode; we'll
  catch this and raise a ValueError.

None of these edge cases should be a problem for Pysa, and no existing
call sites are impacted since they all use nonlazy mode, so these
tradeoffs seem acceptable in order to ship a major perf improvement
without overinvesting in code whose lifetime is (we hope) short. Whenever
we make a new, better-typed query API we should account for these edge
cases better.

Reviewed By: alexkassil

Differential Revision: D37659705

fbshipit-source-id: 4b54c97316b4b6f3a77a9c5f8ce165bf660d1f9d