Lazy server for pysa: wire up the new server arg
Summary:
This commit uses the new `skip_initial_type_check` option in the
server arguments to decide whether we should fully populate the
errors or just boot up the server with an unpopulated lazy
environment.

It is part of a stack to provide users of the server who do not
require a full type check with an option that bypasses the check;
we think this will speed up pysa model generation by something
like 15 minutes (it does on devserver), which will be a big win
for CI and an even bigger win for ourselves whenever we need to
debug the model generation.

Reviewed By: grievejia

Differential Revision: D37654353

fbshipit-source-id: e3f1c8a942d4b34c3aad4983d289d2b0de9f061c