Lazy server for pysa: Pass skip_initial_type_check to server
Summary:
This commit is part of adding the ability to start a "lazy" server
that does not do an initial full check, which will immediately help
pysa and should also enable other use cases that don't need a full-project
type check in order to get some static information from pyre.

In this commit, we pass the new `--skip-initial-type-check` flag to
the server via the arguments json file.

Reviewed By: grievejia

Differential Revision: D37656977

fbshipit-source-id: 8e56811000886ad84505c58f2ca0099ad6d69653