Move coverage gap message to server side
Summary:
### Context:
We want to surface expression level coverage gaps through VSCode highlighting.

### Problem:
In previous diffs, we surfaced the highlighting with messages, but in order to change the messages we must modify the client side.

### Solution:
The message should be made by the server, as the client should just blindly print. Hence, we match the reason and return instead a message. We modify the existing reason so that it is now a string.

### Stack diff:
- Surface highlighting
- Move message to server side (this diff)
- Split `TypeIsAny` into parameter and local cases (next)
- Identify more coverage gaps with callables

Reviewed By: grievejia

Differential Revision: D37647025

fbshipit-source-id: df6937a28bde09d17be50aaa8df3e0ad0e79c603