always log telemetry on `create`
Summary:
In an earlier code review there was a suggestion not to separate
`create` and `create_for_production`.

This sounds good to me, except I want to check one thing: if,
in tests, we don't call `Statistics.GlobalState.initialize` that means we
won't do any external logging?

I just want to be sure we don't start logging perf on all our
unit tests, which is why I shied away from doing this before!

Reviewed By: grievejia

Differential Revision: D37637176

fbshipit-source-id: d419fe69d5e26234484b47a817edf01078ca6f35