Add tests illustrating where Pyre fails to handle NoReturn
Summary:
The purpose of `typing.NoReturn` is to allow type checkers to
understand when a function call halts normal control flow.

Currently Pyre's UninitializedLocalCheck relies exclusively
on the CFG (which is not type-aware) for this, and as a result
we cannot handle these cases.

This commit adds test cases illustrating the problem; to fix
it we will either need
- a type-aware version of the CFG computation (this is probably
  not an easy solution)
- logic that allows us to inspect the type of expressions and
  produce Bottom. This is probably the way to go, although
  it might increase compute time quite a bit; if that becomes
  a problem we might need to find a way to combine uninitialized
  local check with type check so that we don't need to resolve
  expressions twice, which would be a bigger refactor.

Reviewed By: grievejia

Differential Revision: D37559949

fbshipit-source-id: 61389a45b211679a64842c317fa968933f802aac