Differentiate parameters and other expressions for the any coverage gap
Summary:
### Context:
Currently, coverage gaps due to an expression having an `Any` type surface the same message, regardless of whether it is a parameter or not.

### Problem:
We want a different message for parameters that are of the `Any` type, to be more informative to the user.

### Solution:
We surface a different message for expressions that contain `$parameter$` and continue to surface a more generic message for other expressions.

Reviewed By: grievejia

Differential Revision: D37788374

fbshipit-source-id: cfe10f808c212c4dfec4860af06b3118e2e7fe7a