Add socket path to the info output
Summary:
One of the goals of adding `pyre info` is to enable us to work
with the server connection directly when testing out features that
aren't yet ready to expose to end users.

To do that, we want to include the socket path in the output which
would allow testing scripts to create their own connections. Our
plan for the short run to test unsaved changes is to use this.

Reviewed By: grievejia

Differential Revision: D37796003

fbshipit-source-id: 6ca7bec22d4a322c9df1c01c1b2b0a21caedb384