Unsaved changes: add new request type for testing overlays
Summary:
After some debate about pros and cons of various approaches, I've decided
to push forward with a very simple initial implementation for exposing overlays
in the server API, which is to add a new request type that, for now, will
consume full-code updates.

This approach strikes a good balance between quick iteration now and forward
compatibility:
- It avoids mucking up the query type, which I don't want to do because
  updates are fundamentally different (for example, queries only use read-only
  data but updates require read-write access to the overlaid environment)
- We will amost certainly want to iterate on the API when we hook up to persistent
  code, and having it seprate from query makes iteration a bit easier
- If we wind up needing smart batching of updates, we can iterate on a separate
  request type more easily

The plan is to not expose the new functionality at all over the cli for now;
we can do that later, but in the interim we can directly connect to the server
socket from seperate python code in order to test.

Reviewed By: pradeep90

Differential Revision: D37796004

fbshipit-source-id: 582367f5be5f5ec7b340c5979f835d7b1c39431c