cut implicit namespace modules: change module tracker
Summary:
**About this stack**

As we planned out lazy module tracking, a key risk we identified is that the
existing Pyre logic of treating any directory with python files as descendants
as a valid implicit namespace module creates a major bottleneck - to lazily
identify a directory as *not* valid could require a huge, expensive recursive
scan (imagine a large C++ codebase with no python code as a folder in a repo
with lots of python code).

We decided that it makes sense for Pyre to only support a restricted notion
of implicit namespace modules, where only directories with python files as
direct children are considered valid to import in Pyre. This is technically
only a subset of valid Python code since CPython allows you to import
arbitrary directories, but making use of namespace modules with no explicit
modules as children is almost never useful - after some tiny fixes that actually
aren't connected directly to this change but rather to `resolve_exports`
behavior we actually saw *no uses whatsoever* of such imports in distillery,
which is a strong signal that this is a good tradeoff.

**About this diff**

The rest of the stack prepared Pyre to be ready for ModuleTracker to
start only tracking direct children rather than recursive children in
the `implicit_submodules` field of the `Layouts.t` record.

This diff actually makes the cutover.

Reviewed By: grievejia

Differential Revision: D37771613

fbshipit-source-id: 88990c8d70d810d30cb16021c17b21e21cec3730