cut implicit namespace modules: Adapt `resolve_exports` code
Summary:
**About this stack**

The biggest challenge in lazy module tracking is handling implicit
namespace modules. If we want to prevent recursively searching an
entire directory tree (which could be huge - for example a C++
directory embedded in the same repo as python code but with no
python code under that tree), we cannot preserve the existing
behavior of treating any directory with python files under it
as a valid namespace module.

Instead, our plan is to only allow a single level of implicit
namespace modules so that a path is importable only if it either
- corresponds to an explicit module, or
- corresponds to a directory that *directly* contains a python
  file as a child.

This represents a small regression relative to Python, which happily
accepts any directory as a namespace package root, but in practice
it isn't a meaningful regression because namepsace packages with no
immediate explicit children are almost never useful (they cannot possibly
contain useful type information or code - the only thing one can do with
them is import them and use a handful of dunder attributes).

**About this diff**

This diff prepares for refactoring the code to recognize only one layer of
implicit namespace modules, by modifying the `resolve_exports` function to
not assume that all ancestors of a reference's module are themselves valid
modules.

Reviewed By: grievejia

Differential Revision: D37758546

fbshipit-source-id: 1b16b3f27b087c5fed5b4f0281bf97cba3a1074a