cut implicit namespace modules: adapt typeCheck.ml
Summary:
**About the stack**

We've identified the biggest risk to lazy module tracking as being
the way that pyre has historically defined implicit namespace modules:
any directory with a python file as a descendent is treated as a valid
implicit module.

This is a problem for lazy checks because there is no way, short of
full recursive enumeration, to know for sure when a large directory
tree does *not* contain python code - for example if a project has
a huge directory of only C++ sources, pyre could block for a long time
while scaning the full tree.

Instead, we want to only allow the direct import of namespace modules
when they have python files as *direct* children. This approach is
more conservative than CPython, but it's okay for our static analyzer
to reject some valid code in the name of consistency + speed and code
that makes use of implicit grandparent modules is almost never useful
(the only thing you could do with such a module is check some dunder
attributes - it cannot ever have useful functions or type information).

**About this diff**

In order to remove support for arbitrary-depth implicit namespace
modules, we need to have typeCheck.ml's logic for resolving a
`Name` node in the Ast to short-circuit the recursion that currently
jumps all the way to the left-most name.

Instead, we first ask whether the Name is nothing but a reference
(this is not always true, for example `foo().bar` is a `Name` but
the base is a `Call` and it doesn't correspond to a reference).

If so, we then ask `resolve_exports` whtehr name is a valid top-level module
attribute (directly defined, imported, or via a placeholder stub)
and if so we short-circuit by using `forward_reference`.

The resulting behavior produces only about 15 new type check errors on
distillery, all of them on re-exports where we already know Pyre sometimes
struggles due to `legacy_resolve_exports`. These errors are fixed by
using direct imports in D37767925.

Reviewed By: grievejia, pradeep90

Differential Revision: D37768410

fbshipit-source-id: 86ce38454f1fbb42d63a06a863c2ec65a2f24ef7