Extract ResolvedReference.as_module_toplevel_reference
Summary:
This funciton attempts to convert a `ResolvedReference` coming from
`resolve_exports` into an ordinary `Reference.t`, but only in the case
where the resolved reference points to some top-level module attribute.

If the resolved reference points to a nested attribute, it returns None;
the reason is that this is what works best for the current logic that
recursively resolves `Expression.Name` values in `typeCheck.ml`.

Reviewed By: grievejia

Differential Revision: D37866933

fbshipit-source-id: f50df48c2983c1bea9381c48c3e7b3599ef82045