Lazy ModuleTracker: verify namespace package behavior
Summary:
In CPython, a concrete module for a package (either a normal python file
or an `__init__.py` file will force that name to be resolved explicitly,
and will "mask" any implicit namespace package.

So, for example, if there is `/a.py` on any part of the PYTHONPATH, then
a file `/a/b.py` in some other search root will not be visible. Similarly
a `c/__init__.py` will prevent some other root's `c/d.py` from being
visible.

Pyre does not model this because it makes incremental behavior easier,
and in practice this kind of import issue will be trivially exposed by
unit tests.

It will be significantly easier to roll out lazy tracking
if we preserve the existing behavior at first, so this commit adds a
test explaining our behavior and why.

Reviewed By: grievejia

Differential Revision: D37913294

fbshipit-source-id: ed6b80728aa45d20f5dbe6f3d199916c3d0aadd2