Make sure we print the correct buck command on buck failures
Summary:
We print a Log.debug message before running buck, but debug logs are off
by default so it really isn't every helpful.

We also log at warning level if buck fails, but we incorrectly print the
command as `buck` always, even when using buck2. This commit fixes the
problem, by representing the buck command as part of the error.

It would have worked to instead save the buck version in the error, but I
think this is better because it allows me to declare the buck command as
a variable (`buck1` or `buck2`) and use that to ensure that I'm passing
the same command to both `LwtSubprocess.run` and to `on_completion`; if
I used an int it would actually be easier to introduce bugs.

Reviewed By: grievejia

Differential Revision: D37939515

fbshipit-source-id: 7227a28279b7633c860fc34f8f156ae8c2482421