Always pass qualifiers to run on for populate functions
Summary:
I want to make it easier to run a pyre check on just one module;
this is a key blocker for handling edge cases (both logical and performance)
that show up in production when we aren't able to get a repro with a small
example.

My use case is that module tracker changes seem to have broken re-export
resolution in some cases, but in the typical case it works just fine. As a
result it's not obvious how to create a small example of the bug, and I
need debug logging in production to figure out what causes the problem.

This commit doesn't add the ability to run a one-module check per-se, but
it reorganizes the code so that it's possible to make a one-line change
to `checkCommand` where you hardcode the module.

Reviewed By: grievejia

Differential Revision: D38021395

fbshipit-source-id: e22efe8f4da4bdc45861ca503fb57601eb761d5b