Back out: cut implicit namespace modules: change module tracker
Summary:
This change caused edge cases where we fail to properly resolve recursive imports.

The problem is due to a bug in `legacy_resolve_exports` where it still assumes that all implicit modules exist; this
bug was hard to catch because it only shows up in special cases (which were almost nonexistent on distillery) and
the function always falls back.

I understand the issue and will work to fix it, but that might cause more type errors (because we know some code
is actually relying on incorrect resolution) which means it may take a few days.

In the meantime it may be necessary to temporarily revert because the bad case - which didn't happen in distillery
- is showing up more often in some fbcode projects.

Reviewed By: lolpack

Differential Revision: D38050604

fbshipit-source-id: a829e0d94dd610ccddef14b009814df2c17569fd