Support more restrictive types on self or class parameters when overriding
Summary:
We get an error when overriding a method that uses `Self`, because `Self` binds to a more restrictive type in the overriding method. For example:
```
from typing import Self

class Shape:
  def __init__(self, scale: float = 0.0) -> None:
    self.scale = scale

  def set_scale(self: Self, scale: float) -> Self:
    self.scale = scale
    return self

class Circle(Shape):
  def set_scale(self: Self, scale: float) -> Self:
    self.scale = scale + 1.0
    return self
```
Here, `Self` binds to `Shape` in `Shape.set_scale()`, but it binds to `Circle` in `Circle.set_scale()`. Overriding method parameters normally must be less restrictive than those of the overridden method, so we spit out an error in this case, even though we know `self` will always be a subtype of `Circle` in `Circle.set_scale()`.

We fix this by actually solving a more general case: ignore the covariance requirement on the self or class parameter of a method. Thus, annotating the self or class parameter with the actual type is now supported as well. For example:
```
class Shape:
  def __init__(self: Shape, scale: float = 0.0) -> None:
    self.scale = scale

  def set_scale(self: Shape, scale: float) -> Shape:
    self.scale = scale
    return self

class Circle(Shape):
  def set_scale(self: Circle, scale: float) -> Circle:
    self.scale = scale + 1.0
    return self
```
This behavior is supported in [Mypy](https://mypy-play.net/?mypy=latest&python=3.10&gist=d44f9d2b7147050a035cacaa210ee3b6) and Pyright as well.

Reviewed By: pradeep90

Differential Revision: D38009625

fbshipit-source-id: c4fc5c1eb2fca2553c12f05a209e632f3ddb985d