Get rid of `new`-prefix on commands [1/3]: duplicate on server
Summary:
We do not need the newcommand names anymore, they are a historical
artifact from rollout but at this point are just confusing.

I'd like to change the names. This requires a client-server version bump which means it will have to be spaced out quite a bit (I'll use a calendar reminder to bump the client after 6 months, because we know the binary can take a *long* time to roll out to all projects)

Reviewed By: grievejia

Differential Revision: D37774069

fbshipit-source-id: ce50a2a01376b7fe12a79ded728d159cea100b69