Implement Client Call to Overlays Request
Summary:
**Context**: OverlaysRequest has been implemented in the Ocaml server. At this stage, we are able to react to updates in the form of `textDocument/didChange` events. We will handle those events by forwarding an overlaysRequest to the server to update the internal environment stack in the server.

**Considerations**
**1**. At the moment, we do not consume the `TypeErrors` returned to us back from the server as a response to the OverlaysRequest.  This is not done in this diff for three reasons:
a. We don't know if we want to display `TypeErrors` in real time: we could simply load the type errors when a user saves. (Pradeep/Jia could you comment on this?)
b. Display type errors could be extra throwaway work if we were to use a different daemon for the Code Navigation Server so I was hesitant to implement it again.
c. Nontrivial work: To implement this logic wouldn't be difficult, but could prove tedious because we have functions in `persistent.py` that do almost the same thing except that the responses are slightly different: the responses for other requests are wrapped in a format denoting the response is a query, which is different from how it is represented as a response to `OverlaysRequest`. I would need to a small refactor here to avoid duplicating the code.
**2**. I did a slight refactor to make query requests more generic to not just include queries, but also `OverlayRequests`. That involved renaming a few methods and reusing the logic for handling queries to also apply to OverlayRequests.
**3**. We need to decide on a format for overlay_ids. At the moment, I'm just making it the name of the file, but it could also include the PID or a GUID that is propagated through to the server to provide distributed tracing.

Please let me know if you feel this is a feature worth implementing at this stage/ later stages.

**Potentially Encouraging News**

Steven & I previously operated off the assumption that `didChange` events would be sent for every single keystroke, which could be a problem because every letter typed would result in an `overlayRequest`. However, I observed that in reality, it seems like these `didChange` events are sent every second/few seconds rather than every keystroke, which means that potentially our performance problems might not be *as* bad as if every keystroke registered an event.

Note: didChange is only debounced because I specified the `TextDocumentSyncMode` to be `Full` rather than `Incremental` - `Incremental` would result in no debouncing but it would register every keystroke as an event. Currently the changes to the server are assuming that the sync mode is `Full`

**Next Steps**

1. Correctness testing: seeing if `pyre incremental` and a simulated overlay update give the same result for actual production diffs.
2. Combine Unsaved Changes w/ Go-To-Definition: Go to definition currently doesn't utilize unsaved changes: I'll need to make a change to the implementation of Go to Definition to use unsaved changes.
3. Performance tracking: Logging events to see the bottlenecks in overlay type checking. Also, testing these changes on a larger Instagram file could give us qualitative performance feedback.
4. Batching Requests to reduce load on the server.

Reviewed By: grievejia

Differential Revision: D37901111

fbshipit-source-id: f6980ca5642182dc7c3b6b159e6f6a3dfb789799